+++ CHANGELOG.md
@@ -4,6 +4,16 @@
 
 Nothing yet.
 
+## 2.0.0 - Sat Jun 27 2015
+
+### Backwards-Compatibility Breaks
+
+* Objects namespace is now ClassesAndObjects
+
+### New
+
+* Classes\Objects\InvokeMethod::onString()
+
 ## 1.0.0 - Sat Jun 27 2015
 
 Initial release.